backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-core==2.16.2
typing-extensions==4.9.0
anyio==4.2.0
starlette==0.36.3
python-multipart==0.0.6
pyjwt==2.8.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1